% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_spss_read.R
\name{gesis_spss_read}
\alias{gesis_spss_read}
\title{Read in a GESIS SPSS file.}
\usage{
gesis_spss_read(path = NULL, zacat_id = NULL, conversion = "factor")
}
\arguments{
\item{path}{An object to be converted to character}

\item{zacat_id}{Default to \code{NULL}. Not used in this instance
of the method.}

\item{conversion}{Defaults to \code{'factor'}.}
}
\description{
Read in a GESIS SPSS file.
}
\examples{
\dontrun{
##use your own file:
gesis_spss_read(    path = file.path ( 'not_included', 'example.sav'),
                    zacat_id = NULL )
}
}
